Source code release & all selfbot assets
Last updated on 11/1/2021 at 10:15 AM, Phantom was a state-of-the-art Discord selfbot that allowed users to execute custom code on behalf of themselves by running simple chat commands. The bot featured over 750 commands in the free tier and 1,000 in the premium tier. Complete with a responsive backend api written in PHP, a secure authentication system in PHP I still use for my other projects, and 25k+ lines of code, Phantom was a revolutionary selfbot that changed the game forever.

The main selling point for Phantom was to increase competition with other public selfbots and reduce their prices. A popular selfbot, Nighty, was previously priced at 20 USD. After the release of Phantom, they had to lower their prices in order to compete, as Phantom included all the features of their selfbot but completely free of charge. However, I stopped developing Phantom not much longer than a year after it's release. Nevertheless, Nighty's price today (6/18/2023) remains at 9 USD.

NOTE: All the API endpoints the selfbot used (except for the authentication system in PHP) are now deprecated and no longer respond.
However, assets such as extra files, resources, libraries, sounds, and other cloud-hosted files are still available and will respond.